requirements.txt
selenium>=4.18
pytest>=8.0
pytest-html>=4.0
webdriver-manager>=4.0
python-dotenv>=1.0
ruff>=0.3.0
